moves search component into customHook - updates toastID to ref to dismiss
